+++ src/Forms/UI/views/fields/text.blade.php
@@ -18,4 +18,8 @@
             ->merge($getCustomAttributes())
             ->merge($modelBinding)"
     />
+
+    @if($isHiveEnabled())
+        @include('chief-hive::suggest', ['payload' => $getHivePayload($locale ?? null, $this->getId() ? $this : null)])
+    @endif
 </x-chief::form.input.prepend-append>