+++ src/Plugins/Seo/App/routes/chief-admin-routes.php
@@ -3,4 +3,4 @@
 use Illuminate\Support\Facades\Route;
 
 Route::get('seo', [\Thinktomorrow\Chief\Plugins\Seo\App\Controllers\SeoController::class, 'index'])->name('chief.seo.index');
-Route::get('seo/alt', [\Thinktomorrow\Chief\Plugins\Seo\App\Controllers\SeoController::class, 'altIndex'])->name('chief.seo.alt.index');
+Route::get('seo/assets', [\Thinktomorrow\Chief\Plugins\Seo\App\Controllers\SeoController::class, 'assetsIndex'])->name('chief.seo.assets.index');