+++ src/Plugins/Seo/UI/views/index.blade.php
@@ -3,7 +3,7 @@
         <div class="-my-4 divide-y divide-grey-100">
             <div class="flex items-center justify-between py-3">
                 <a
-                    href="{{ route('chief.seo.alt.index') }}"
+                    href="{{ route('chief.seo.assets.index') }}"
                     title="Alt"
                     class="body-dark font-medium hover:underline"
                 >
@@ -11,7 +11,7 @@
                 </a>
 
                 <x-chief::button
-                    href="{{ route('chief.seo.alt.index') }}"
+                    href="{{ route('chief.seo.assets.index') }}"
                     title="Alt"
                     size="sm"
                     variant="grey"